requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b0
click==7.1.2
Django==2.2.10
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.8.1
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
